Added the Tapered Information
ID and OD max do not refine the table for the multi-extrusion.

In the tapered UI, the values are not reset once deselected.

Once deselected, both values are reset if they are in the stack, this causes an issue with observe event since it observes two events changing and grep will match twice.