Added show and hide puttons.
I added show and hide buttons for checkboxes. I added a reset button for all parameters.

I also fixed the grep for single extrusion so "Length (in)" does not mismatch for tip land length of die land length. I did this by using "^" so the parameter must tart with it.